requirements.txt
# Django - Django REST
Django==4.1.5
django-cors-headers==3.13.0
django-redis==5.2.0
djangorestframework==3.14.0
django-prometheus==2.2.0
django-dotenv==1.4.2
gunicorn==19.10.0
prometheus-client==0.15.0


# JWT Token
PyJWT==2.6.0


# Schedule tasks
schedule==1.1.0


# Database
mysqlclient==2.1.1
redis==4.4.2


# Python utils
asgiref==3.6.0
python-dateutil==2.8.2
pytz==2022.7
sqlparse==0.4.1


# Slack API
slack_sdk==3.0.0rc1

# Env variables management
envparse==0.2.0

